data entry

+++ src/main/java/com/jica/sdg/repository/EntryNsaIndicatorRepository.java
@@ -9,7 +9,7 @@ import javax.transaction.Transactional;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.repository.query.Param;
 
-public interface EntryNsaIndicatorRepository extends CrudRepository<EntryNsaIndicator, Long> {
+public interface EntryNsaIndicatorRepository extends CrudRepository<EntryNsaIndicator, Integer> {
 
     
 }
